refactor: extract transformCode to output function
