+++ packages/jsx2mp-loader/src/output.js
@@ -1,7 +1,51 @@
 const { writeJSONSync, writeFileSync, existsSync, mkdirpSync } = require('fs-extra');
 const { extname, dirname, join } = require('path');
+const { transformSync } = require('@babel/core');
 const { minify, minifyJS, minifyCSS, minifyXML } = require('./utils/minifyCode');
 const addSourceMap = require('./utils/addSourceMap');
+
+function transformCode(rawContent, mode, externalPlugins = [], externalPreset = []) {
+  const presets = [].concat(externalPreset);
+  const plugins = externalPlugins.concat([
+    require('@babel/plugin-proposal-export-default-from'), // for support of export defualt
+    [
+      require('babel-plugin-transform-define'),
+      {
+        'process.env.NODE_ENV': mode === 'build' ? 'production' : 'development',
+      }
+    ],
+    [
+      require('babel-plugin-minify-dead-code-elimination'),
+      {
+        optimizeRawSize: true,
+        keepFnName: true
+      }
+    ],
+  ]);
+
+
+  const babelParserOption = {
+    plugins: [
+      'classProperties',
+      'jsx',
+      'flow',
+      'flowComment',
+      'trailingFunctionCommas',
+      'asyncFunctions',
+      'exponentiationOperator',
+      'asyncGenerators',
+      'objectRestSpread',
+      ['decorators', { decoratorsBeforeExport: false }],
+      'dynamicImport',
+    ], // support all plugins
+  };
+
+  return transformSync(rawContent, {
+    presets,
+    plugins,
+    parserOpts: babelParserOption
+  });
+}
 
 /**
  * Process and write file
@@ -10,13 +54,22 @@ const addSourceMap = require('./utils/addSourceMap');
  * @param {object} options
  */
 function output(content, raw, options) {
-  const { mode, outputPath } = options;
+  const { mode, outputPath, externalPlugins = [] } = options;
   let { code, config, json, css, map, template, assets } = content;
   if (mode === 'build') {
-    // Minify code
-    code = minifyJS(code);
+    // Compile ES6 => ES5 and minify code
+    code = minifyJS(transformCode(code,
+      mode,
+      externalPlugins.concat([require('@babel/plugin-proposal-class-properties')]),
+      [require('@babel/preset-env')]
+      ).code);
     if (config) {
-      config = minifyJS(config);
+      // Compile ES6 => ES5 and minify code
+      config = minifyJS(transformCode(config,
+        mode,
+        externalPlugins.concat([require('@babel/plugin-proposal-class-properties')]),
+        [require('@babel/preset-env')]
+      ).code);
     }
     if (css) {
       css = minifyCSS(css);
@@ -25,6 +78,7 @@ function output(content, raw, options) {
       template = minifyXML(template);
     }
   } else {
+    code = transformCode(code, mode, externalPlugins).code;
     // Add source map
     if (map) {
       code = addSourceMap(code, raw, map);
